fix(app): share worktree/notification event handling across layouts
The worktree.ready/failed listener that releases the new-worktree prompt gate lived only in the legacy layout, so under the new layout designs a New Workspace session would create the worktree but never receive the ready signal, leaving the first prompt stalled indefinitely with no title. Extract useSDKNotificationToasts into a shared useNotificationToasts hook and mount it from both layouts so the gate is released and permission/question toasts fire regardless of active design. Fixes #36731.
